Made some optimizations to the share watch experience in the web application.
