+++ web/src/types/index.ts
@@ -131,4 +131,6 @@ export type {
   CreateWatchRoomResponseType,
   WatchRoomResponseType,
   JoinWatchRoomResponseType,
+  WatchRoomPlaybackStateType,
+  WatchRoomServerEventType,
 } from "./watch-rooms";
